+++ src/input/input_field.rs
@@ -2,7 +2,7 @@ use crate::cycle::CycleRecoveryStrategy;
 use crate::ingredient::{fmt_index, Ingredient};
 use crate::input::Configuration;
 use crate::local_state::QueryOrigin;
-use crate::storage::IngredientIndex;
+use crate::zalsa::IngredientIndex;
 use crate::{Database, DatabaseKeyIndex, Id, Revision};
 use std::fmt;
